Add DatadogTraceProcessor
This processor adds Datadog tracing information to the event dictionary if the
ddtrace library is installed. If the library is not installed it is a no-op
processor. If it is installed it will conditionally add fields to the event
dictionary if they are present. Unlike the reference implementation (located
here:
https://docs.datadoghq.com/tracing/other_telemetry/connect_logs_and_traces/python/#no-standard-library-logging)
if the library cannot find a current span or a config value is not defined the
processor will not add the key to the event dictionary.