Add DatadogTraceProcessor to the JSONLogToStdout config
This is the config used in deployed environments where the Datadog agent is
available. The DevLogToStdout is really only used in local development where
the agent is more likely to not be available and is even less likely to be
gathering logs. Since the trace and span IDs are used to correlate logs to
traces it doesn't make a lot of sense to add it to DevLogToStdout but should
definitely be added to JSONLogToStdout. Given that the processor will no-op if
the ddtrace library is not present this seems like a reasonable decision.